moved get_ip_from_interface, get_linux_uptime func to networking.py
