+++ agents/dns/config.ini
@@ -1,7 +1,7 @@
 ; Configuration applied to the DNS agent.
 [server]
 ; Name of the server (not to be confused with hostname)
-name=
+name=dns
 port=53
 ; Must be a single value or a comma-separated list of processes
 processes=named
